requirements (copy).txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.12.1
backports.zoneinfo==0.2.1
bidict==0.23.0
blinker==1.7.0
boto3==1.34.41
botocore==1.34.41
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
fastjsonschema==2.19.1
flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flask-sqlalchemy==3.1.1
frozenlist==1.4.1
fyers-apiv2==2.0.7
fyers-apiv3==3.0.7
greenlet==3.0.3
h11==0.14.0
idna==2.10
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
logzero==1.7.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.5
NorenRestApiPy==0.0.22
numpy==1.24.4
packaging==23.2
pandas==2.0.3
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
simple-websocket==1.0.0
sip==6.8.1
six==1.16.0
smartapi-python==1.4.1
SQLAlchemy==2.0.25
tomli==2.0.1
tornado==6.1
typing-extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
websocket-client==1.6.1
websockets==8.1
werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
